==> 04_assignmentfile_markdownreport.Rmd <==
---
title: "GW Assignment - Biden Transition"
author: "YOUR NAME HERE"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(janitor)
library(scales)
library(gt)
library(kableExtra)
options(dplyr.summarise.inform = FALSE)

#run the script for step 03 to update the datasets to capture the current live verions
#and load all the new dataset objects into the environment (you'll see them populate on the upper right window)
source("03_compare_current_with_previous.R")


```

**Note: use either kable/kableExtra or gt tables to format your final tables for the questions below. You'll likely find it easier to figure out the data parts first, then once that's done add the table formatting.**

# AGENCY TEAMS

Come up with the necessary R code to return the following for the agency review teams.

### *New Names*

Below write code to show the new names added to the agency review team lists since the prior data provided.  
  

```{r, echo=FALSE}



 

```


### *Aggregate Team Changes*

Add data to show the total number of people appointed to each agency team, along with change since last time reflecting the number of new people added. Omit agencies with no change at all.  
  

```{r, echo=FALSE}





```


### *Largest and Smallest Overall*

Show the top 10 **largest** agency review teams as of today:

```{r, echo=FALSE}






```

Show the top **smallest** agency review teams as of today - which we'll define here as less than five members:

```{r, echo=FALSE}




 

```



# WHITE HOUSE SENIOR STAFF

Come up with the necessary R code to return the following for the WH senior staff.

### *New Names*

Below write code to show the new names added to the senior staff lists since the prior data provided.  
  

```{r, echo=FALSE}



 

```


### *Total Number of Current vs. Previous*

Add code to show the total number of people currently named to the WH senior staff, vs. the previous total number.  
  

```{r, echo=FALSE}





```
